Ads more helpful error messages.
Instead of having error-reporting code strewn about in printf + exit
statements, all error reporting code will reside in error.c.

`err_at_tok` works by taking a filepath and a token and printing the
offending code line, with a red highlight over the passed token.

A filepath has to be passed because source files are not read to memory
but streamed on a need to read basis into a double buffer. The error
reporting function has to open the offending source file and seek the
requested line.